root checks and ntp

+++ install_tendermint.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+if [ `whoami` == "root" ];
+then
+	echo "You should not run this script as root"
+	exit 1
+fi
+
 USER=`whoami`
 source /home/$USER/.bashrc
 rm -rf $GOPATH/src/github.com/tendermint/tendermint
